+++ src/main/kotlin/io/openfuture/chain/entity/transaction/BaseTransaction.kt
@@ -1,10 +1,10 @@
 package io.openfuture.chain.entity.transaction
 
 import com.fasterxml.jackson.annotation.JsonIgnore
+import io.openfuture.chain.crypto.signature.SignatureManager
 import io.openfuture.chain.crypto.util.HashUtils
-import io.openfuture.chain.entity.base.BaseModel
-import io.openfuture.chain.entity.Block
 import io.openfuture.chain.entity.MainBlock
+import io.openfuture.chain.entity.base.BaseModel
 import javax.persistence.*
 
 @Entity
@@ -30,15 +30,23 @@ abstract class BaseTransaction(
     @Column(name = "sender_address", nullable = false)
     var senderAddress: String,
 
-    @Column(name = "sender_signature", nullable = false)
-    var senderSignature: String,
-
     @Column(name = "hash", nullable = false)
     var hash: String,
+
+    @Column(name = "sender_signature", nullable = false)
+    var senderSignature: String? = null,
 
     @JsonIgnore
     @ManyToOne
     @JoinColumn(name = "block_id", nullable = true)
     var block: MainBlock? = null
 
-) : BaseModel()
+) : BaseModel() {
+
+    @Suppress("UNCHECKED_CAST")
+    fun <T : BaseTransaction> sign(privateKey: ByteArray): T {
+        this.senderSignature = SignatureManager.sign(HashUtils.fromHexString(hash), privateKey)
+        return this as T
+    }
+
+}