+++ src/main/kotlin/io/openfuture/chain/entity/transaction/TransferTransaction.kt
@@ -1,5 +1,6 @@
 package io.openfuture.chain.entity.transaction
 
+import io.openfuture.chain.crypto.util.HashUtils
 import io.openfuture.chain.domain.rpc.transaction.TransferTransactionRequest
 import io.openfuture.chain.domain.transaction.TransferTransactionDto
 import io.openfuture.chain.entity.MainBlock
@@ -15,10 +16,13 @@ class TransferTransaction(
     recipientAddress: String,
     senderKey: String,
     senderAddress: String,
-    senderSignature: String,
-    hash: String,
+    hash: String? = null,
+    senderSignature: String? = null,
     block: MainBlock? = null
-) : BaseTransaction(timestamp, amount, fee, recipientAddress, senderKey, senderAddress, senderSignature, hash, block) {
+) : BaseTransaction(timestamp, amount, fee, recipientAddress, senderKey, senderAddress,
+    hash ?: HashUtils.toHexString(
+        HashUtils.sha256((senderAddress + recipientAddress + timestamp + amount + fee).toByteArray())),
+    senderSignature, block) {
 
     companion object {
         fun of(dto: TransferTransactionDto): TransferTransaction = TransferTransaction(
@@ -28,8 +32,8 @@ class TransferTransaction(
             dto.recipientAddress,
             dto.senderKey,
             dto.senderAddress,
-            dto.senderSignature,
-            dto.hash
+            dto.hash,
+            dto.senderSignature
         )
 
         fun of(timestamp: Long, request: TransferTransactionRequest): TransferTransaction = TransferTransaction(
@@ -39,8 +43,8 @@ class TransferTransaction(
             request.recipientAddress!!,
             request.senderKey!!,
             request.senderAddress!!,
-            request.senderSignature!!,
-            request.getHash()
+            request.getHash(),
+            request.senderSignature!!
         )
     }