+++ src/main/kotlin/io/openfuture/chain/entity/transaction/VoteTransaction.kt
@@ -1,5 +1,6 @@
 package io.openfuture.chain.entity.transaction
 
+import io.openfuture.chain.crypto.util.HashUtils
 import io.openfuture.chain.domain.rpc.transaction.VoteTransactionRequest
 import io.openfuture.chain.domain.transaction.VoteTransactionDto
 import io.openfuture.chain.entity.MainBlock
@@ -18,8 +19,6 @@ class VoteTransaction(
     recipientAddress: String,
     senderKey: String,
     senderAddress: String,
-    senderSignature: String,
-    hash: String,
 
     @Column(name = "vote_type_id", nullable = false)
     private var voteTypeId: Int,
@@ -30,9 +29,14 @@ class VoteTransaction(
     @Column(name = "delegate_port", nullable = false)
     var delegatePort: Int,
 
+    hash: String? = null,
+    senderSignature: String? = null,
     block: MainBlock? = null
 
-) : BaseTransaction(timestamp, amount, fee, recipientAddress, senderKey, senderAddress, senderSignature, hash, block) {
+) : BaseTransaction(timestamp, amount, fee, recipientAddress, senderKey, senderAddress,
+    hash ?: HashUtils.toHexString(HashUtils.sha256(
+        (senderAddress + recipientAddress + timestamp + amount + fee + voteTypeId + delegateHost + delegatePort).toByteArray())),
+    senderSignature, block) {
 
     companion object {
         fun of(dto: VoteTransactionDto): VoteTransaction = VoteTransaction(
@@ -42,11 +46,11 @@ class VoteTransaction(
             dto.recipientAddress,
             dto.senderKey,
             dto.senderAddress,
-            dto.senderSignature,
-            dto.hash,
             dto.voteType.getId(),
             dto.delegateInfo.networkAddress.host,
-            dto.delegateInfo.networkAddress.port
+            dto.delegateInfo.networkAddress.port,
+            dto.hash,
+            dto.senderSignature
         )
 
         fun of(timestamp: Long, request: VoteTransactionRequest): VoteTransaction = VoteTransaction(
@@ -56,12 +60,11 @@ class VoteTransaction(
             request.recipientAddress!!,
             request.senderKey!!,
             request.senderAddress!!,
-            request.senderSignature!!,
-            request.getHash(),
             request.voteType!!.getId(),
             request.delegateInfo!!.networkAddress.host,
-            request.delegateInfo!!.networkAddress.port
-
+            request.delegateInfo!!.networkAddress.port,
+            request.getHash(),
+            request.senderSignature!!
         )
     }