+++ src/main/kotlin/io/openfuture/chain/service/transaction/DefaultCoinBaseTransactionService.kt
@@ -4,6 +4,7 @@ import io.openfuture.chain.component.node.NodeClock
 import io.openfuture.chain.crypto.key.NodeKeyHolder
 import io.openfuture.chain.crypto.util.HashUtils
 import io.openfuture.chain.entity.transaction.CoinBaseTransaction
+import io.openfuture.chain.property.ConsensusProperties
 import io.openfuture.chain.repository.CoinBaseTransactionRepository
 import io.openfuture.chain.service.CoinBaseTransactionService
 import io.openfuture.chain.service.DelegateService
@@ -15,23 +16,19 @@ class DefaultCoinBaseTransactionService(
     repository: CoinBaseTransactionRepository,
     private val nodeClock: NodeClock,
     private val keyHolder: NodeKeyHolder,
-    private val delegateService: DelegateService
+    private val delegateService: DelegateService,
+    private val consensusProperties: ConsensusProperties
 ) : DefaultBaseTransactionService<CoinBaseTransaction>(repository), CoinBaseTransactionService {
-
-    companion object {
-        private const val COINBASE_ADDRESS = "0x0000000000000000000000000000000000000000"
-    }
-
 
     @Transactional
     override fun save(tx: CoinBaseTransaction): CoinBaseTransaction = repository.save(tx)
 
-    override fun create(): CoinBaseTransaction {
+    override fun create(fees: Double): CoinBaseTransaction {
         val publicKey = HashUtils.toHexString(keyHolder.getPublicKey())
         val delegate = delegateService.getByPublicKey(publicKey)
 
-        return CoinBaseTransaction(nodeClock.networkTime(), 10.0, 0.0, "delegateAddress",
-            publicKey, COINBASE_ADDRESS, "signature", "hash")
+        return CoinBaseTransaction(nodeClock.networkTime(), fees + 10.0, 0.0, delegate.getAddress(),
+            publicKey, consensusProperties.genesisAddress!!).sign(keyHolder.getPrivateKey())
     }
 
 }