+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fees to transactions
 - New type of transaction: CoinBaseTransaction
 - Genesis address to properties
+- Sign transactions
+- Create and save CoinBaseTransaction
 
 ## [0.3.0] - 2018-07-13
 ### Added